requirements.txt
pandas
chromadb
schedule
google-generativeai
streamlit
plotly